Fix SystemTap dtrace warning for smgr probe argument types
Commits ca326e903d and 1f8c504e308 widened the byte-count arguments of
smgr__md__read__done and smgr__md__write__done to "long long int".
SystemTap's dtrace(1) pyparsing grammar rejects that multi-word type and
falls back with a warning (misreported near the previous probe).  Use
"long" instead, which matches ssize_t on our LP64 DTrace platforms and
is already used elsewhere in probes.d.

Author: Andrey Rachitskiy <pl0h0yp1@gmail.com>
Reported-by: Laurenz Albe <laurenz.albe@cybertec.at>
Discussion: https://postgr.es/m/697d3c88568442cd637d8453d129f1bb14bbd2a8.camel@cybertec.at

+++ src/backend/utils/probes.d
@@ -13,6 +13,8 @@
  *
  * NOTE: Do not use system-provided typedefs (e.g. uintptr_t, uint32_t, etc)
  * in probe definitions, as they cause compilation errors on macOS.
+ * Also avoid "long long": SystemTap's dtrace(1) rejects it.  Use "long"
+ * for 64-bit quantities on our LP64 DTrace platforms.
  */
 #define LocalTransactionId unsigned int
 #define LWLockMode int
@@ -83,9 +85,9 @@ provider postgresql {
 	probe twophase__checkpoint__done();
 
 	probe smgr__md__read__start(ForkNumber, BlockNumber, Oid, Oid, Oid, int);
-	probe smgr__md__read__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, long long int, long long int);
+	probe smgr__md__read__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, long, long);
 	probe smgr__md__write__start(ForkNumber, BlockNumber, Oid, Oid, Oid, int);
-	probe smgr__md__write__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, long long int, long long int);
+	probe smgr__md__write__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, long, long);
 
 	probe wal__insert(unsigned char, unsigned char);
 	probe wal__switch();
